requirements.txt
pandas
requests
pytz
python-amazon-sp-api
google-auth
google-auth-oauthlib
google-api-python-client